Simplificação do banco de dados; melhorias estéticas

+++ userscripts/eproc-favoritos-historico/src/tela_processo.ts
@@ -5,7 +5,11 @@ import * as db from './database';
 import { log_error } from './log_error';
 import classes from './tela_processo.module.scss';
 
-type Status = 'PENDING' | 'ACTIVE' | 'INACTIVE' | 'ERROR';
+type Status =
+  | 'PENDING'
+  | { status: 'ACTIVE'; motivo: string }
+  | 'INACTIVE'
+  | 'ERROR';
 
 export async function tela_processo() {
   const elemento_numero = document.getElementById('txtNumProcesso');
@@ -18,13 +22,21 @@ export async function tela_processo() {
     P.isNonEmptyString(numero_formatado),
     'Erro ao obter o número do processo.'
   );
-  const numero = numero_formatado.replace(/[.-]/g, '');
+  const numero = P.check(
+    db.isNumProc,
+    numero_formatado.replace(/[.-]/g, ''),
+    'Erro ao obter número do processo'
+  );
 
   const status = create_store<Status>('PENDING');
 
   try {
     const resultado = await db.verificar_favorito(numero);
-    status.set(resultado ? 'ACTIVE' : 'INACTIVE');
+    status.set(
+      resultado !== undefined
+        ? { status: 'ACTIVE', motivo: resultado.motivo }
+        : 'INACTIVE'
+    );
   } catch (err) {
     status.set('ERROR');
     throw err;
@@ -36,14 +48,17 @@ export async function tela_processo() {
     const current = status.get();
     if (current === 'ERROR' || current === 'PENDING') return;
     try {
-      if (current === 'ACTIVE') {
+      if (typeof current === 'object' && current.status === 'ACTIVE') {
         status.set('PENDING');
         await db.remover_favorito(numero);
         status.set('INACTIVE');
       } else {
         status.set('PENDING');
-        await db.salvar_favorito(numero);
-        status.set('ACTIVE');
+        // TODO
+        const motivo = '';
+        const prioridade = db.Prioridade.MEDIA;
+        await db.salvar_favorito({ numproc: numero, motivo, prioridade });
+        status.set({ status: 'ACTIVE', motivo });
       }
     } catch (err) {
       status.set('ERROR');
@@ -78,7 +93,10 @@ function render_estrela(elemento_numero: HTMLElement) {
   parent.append(div);
   elemento_numero.classList.add('col-auto', 'pr-0');
 
-  const info: Record<Status, { symbol: string; title: string }> = {
+  const info: Record<
+    Status extends infer S ? (S extends { status: infer K } ? K : S) : never,
+    { symbol: string; title: string }
+  > = {
     PENDING: { symbol: 'hourglass_top', title: 'Aguarde' },
     ACTIVE: { symbol: 'star', title: 'Remover dos favoritos' },
     INACTIVE: { symbol: 'star_outline', title: 'Adicionar aos favoritos' },
@@ -88,9 +106,16 @@ function render_estrela(elemento_numero: HTMLElement) {
   return {
     eventTarget,
     update: (status: Status) => {
-      const { symbol, title } = info[status];
+      const { symbol, title } =
+        typeof status === 'string' ? info[status] : info[status.status];
       icon.textContent = symbol;
-      link.title = title;
+      if (typeof status === 'string') {
+        link.title = title;
+      } else if (typeof status === 'object' && status.status === 'ACTIVE') {
+        link.title = status.motivo.trim() || title;
+      } else {
+        throw new Error('Status desconhecido.');
+      }
 
       if (status === 'PENDING') {
         link.classList.add(classes.wait);
@@ -98,7 +123,7 @@ function render_estrela(elemento_numero: HTMLElement) {
         link.classList.remove(classes.wait);
       }
 
-      if (status === 'ACTIVE') {
+      if (typeof status === 'object' && status.status === 'ACTIVE') {
         link.classList.add(classes.added);
       } else {
         link.classList.remove(classes.added);
